Feature/issue 45 separate normal admin privileges

+++ app/routes/c2dns.py
@@ -11,7 +11,13 @@ from app.routes.tags_mapping import create_tags_mapping, delete_tags_mapping
 @app.route('/ThreatKB/c2dns', methods=['GET'])
 @login_required
 def get_all_c2dns():
-    entities = c2dns.C2dns.query.all()
+    entities = c2dns.C2dns.query
+
+    if not current_user.admin:
+        entities = entities.filter_by(owner_user_id=current_user.id)
+
+    entities = entities.all()
+
     return json.dumps([entity.to_dict() for entity in entities])
 
 
@@ -20,6 +26,8 @@ def get_c2dns(id):
     entity = c2dns.C2dns.query.get(id)
     if not entity:
         abort(404)
+    if not current_user.admin and entity.owner_user_id != current_user.id:
+        abort(403)
     return jsonify(entity.to_dict())
 
 
@@ -52,6 +60,8 @@ def update_c2dns(id):
     entity = c2dns.C2dns.query.get(id)
     if not entity:
         abort(404)
+    if not current_user.admin and entity.owner_user_id != current_user.id:
+        abort(403)
     entity = c2dns.C2dns(
         state=request.json['state']['state'] if request.json['state'] and 'state' in request.json['state'] else
         request.json['state'],
@@ -85,6 +95,8 @@ def delete_c2dns(id):
 
     if not entity:
         abort(404)
+    if not current_user.admin and entity.owner_user_id != current_user.id:
+        abort(403)
     db.session.delete(entity)
     db.session.commit()
